Don't drop namespace if it's already gone
In some cases (like manual reloads), old entries in alias map
for wdqs namespaces (mostly categories) may be incorrect, causing
replaceNamespace function to fail on deleting a namespace (and
in consequence the script if it has set -e). This change checks
if namespace exists and doesn't try to delete it when it doesn't.

Bug: T269331

+++ modules/query_service/files/cron/cronUtils.sh
@@ -53,7 +53,11 @@ function replaceNamespace {
 	# Bump nginx to reload config
 	sudo systemctl reload nginx
 	if [ -n "${oldNamespace}" ]; then
-		# Drop old namespace
-		curl -s -X DELETE "${endpoint}${NAMESPACE_URL}${oldNamespace}"
+	  if [ "$(curl -sLI -o /dev/null --fail --write-out "%{http_code}" "${endpoint}${NAMESPACE_URL}${oldNamespace}/sparql")" -eq "200" ]; then
+		  # Drop old namespace
+		  curl -s -X DELETE "${endpoint}${NAMESPACE_URL}${oldNamespace}"
+	  else
+	    echo "Namespace ${oldNamespace} is in the alias file, but does not appear to be present in Blazegraph."
+	  fi
 	fi
 }
